Refactored: Datadog::Error --> Datadog::Core::Error
