+++ lib/ddtrace.rb
@@ -9,7 +9,6 @@ require 'ddtrace/version'
 require 'datadog/tracing'
 require 'ddtrace/pin'
 require 'ddtrace/tracer'
-require 'ddtrace/error'
 require 'ddtrace/quantization/hash'
 require 'ddtrace/quantization/http'
 require 'ddtrace/pipeline'
